Add FilteringSelection struct; Refactoring;
